Combined Oversampling and Under sampling is done
Instead of considering only oversampling or undesampling individually  we are considering both here.
The concept is that we can apply modest amount of oversampling to the minority class, which improves the bias to the minority class examples, whilst we also perform a modest amount of undersampling on the majority class to reduce the biass on the majority class examples.